H.Res. 363 Introduced in House
Requesting the President and directing the Secretary of State, the Secretary of Defense, the Director of the Central Intelligence Agency, and the Attorney General to transmit to the House of Representatives not later than 14 days after the date of the adoption of this resolution documents in the possession of the President and those officials relating to the disclosure of the identity and employment of Ms. Valerie Plame.

Version:  Introduced in House (ih)
Date:     2005-07-14
Congress: 109

Source: https://www.govinfo.gov/content/pkg/BILLS-109hres363ih/xml/BILLS-109hres363ih.xml

Sponsored-By: H001032
Cosponsor-Count: 50