H.Res. 420 Introduced in House
Directing the Attorney General to transmit to the House of Representatives not later than 14 days after the date of the adoption of this resolution documents in the possession of the Attorney General relating to the disclosure of the identity and employment of Ms. Valerie Plame.

Version:  Introduced in House (ih)
Date:     2005-07-29
Congress: 109

Source: https://www.govinfo.gov/content/pkg/BILLS-109hres420ih/xml/BILLS-109hres420ih.xml

Sponsored-By: H001032
Cosponsor-Count: 2