H.Res. 418 Reported in House
Requesting the President to transmit to the House of Representatives not later than 14 days after the date of the adoption of this resolution documents in the possession of the President relating to the disclosure of the identity and employment of Ms. Valerie Plame.

Version:  Reported in House (rh)
Date:     2005-09-21
Congress: 109

Source: https://www.govinfo.gov/content/pkg/BILLS-109hres418rh/xml/BILLS-109hres418rh.xml

Sponsored-By: H001032
Cosponsor-Count: 20