H.Res. 572 Reported in House
Providing for consideration of the resolution (H. Res. 571) expressing the sense of the House of Representatives that the deployment of United States forces in Iraq be terminated immediately and providing for consideration of the concurrent resolution (H. Con. Res. 308) directing the Clerk of the House of Representatives to make a technical correction in the enrollment of H.R. 3058.

Version:  Reported in House (rh)
Date:     2005-11-18
Congress: 109

Source: https://www.govinfo.gov/content/pkg/BILLS-109hres572rh/xml/BILLS-109hres572rh.xml

Sponsored-By: G000550