S. 2144 Introduced in Senate
A bill to authorize the Secretary of the Interior to conduct a study to determine the suitability and feasibility of designating the Solders' Memorial Military Museum located in St. Louis, Missouri, as a unit of the National Park System.

Version:  Introduced in Senate (is)
Date:     2005-12-19
Congress: 109

Source: https://www.govinfo.gov/content/pkg/BILLS-109s2144is/xml/BILLS-109s2144is.xml

Sponsored-By: T000024